+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_help.py
@@ -480,9 +480,12 @@ examples:
   - name: Create a task with the definition from the pipe.
     text: >
         cat task.yaml | az acr task create -n hello-world -r MyRegistry -f - -c /dev/null
-  - name: Create a Linux task from a public GitHub repository which builds the hello-world image without triggers
+  - name: Create a Linux task from a public GitHub repository which builds the hello-world image without triggers.
     text: >
         az acr task create -t hello-world:{{.Run.ID}} -n hello-world -r MyRegistry -c https://github.com/Azure-Samples/acr-build-helloworld-node.git -f Dockerfile --commit-trigger-enabled false --pull-request-trigger-enabled false
+  - name: Create a Linux task from a public GitHub repository which builds the hello-world image without triggers and uses a build argument.
+    text: >
+        az acr task create -t hello-world:{{.Run.ID}} -n hello-world -r MyRegistry -c https://github.com/Azure/acr-builder.git -f Dockerfile --commit-trigger-enabled false --pull-request-trigger-enabled false --arg DOCKER_CLI_BASE_IMAGE=docker:18.03.0-ce-git
   - name: Create a Linux task using a private GitHub repository which builds the hello-world image without triggers on Arm architecture (V7 variant)
     text: >
         az acr task create -t hello-world:{{.Run.ID}} -n hello-world -r MyRegistry -c https://github.com/Azure-Samples/acr-build-helloworld-node.git -f Dockerfile --commit-trigger-enabled false --pull-request-trigger-enabled false --git-access-token 0000000000000000000000000000000000000000 --platform linux/arm/v7