+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_validators.py
@@ -68,6 +68,21 @@ def validate_keyvault(cmd, namespace):
                                           'vaults', 'Microsoft.KeyVault')
 
 
+def validate_proximity_placement_group(cmd, namespace):
+    from msrestazure.tools import parse_resource_id
+
+    if namespace.proximity_placement_group:
+        namespace.proximity_placement_group = _get_resource_id(cmd.cli_ctx, namespace.proximity_placement_group,
+                                                               namespace.resource_group_name,
+                                                               'proximityPlacementGroups', 'Microsoft.Compute')
+
+        parsed = parse_resource_id(namespace.proximity_placement_group)
+        rg, name = parsed['resource_group'], parsed['name']
+
+        if not check_existence(cmd.cli_ctx, name, rg, 'Microsoft.Compute', 'proximityPlacementGroups'):
+            raise CLIError("Proximity Placement Group '{}' does not exist.".format(name))
+
+
 def process_vm_secret_format(cmd, namespace):
     from msrestazure.tools import is_valid_resource_id
     from azure.cli.core._output import (get_output_format, set_output_format)
@@ -1027,6 +1042,8 @@ def process_vm_create_namespace(cmd, namespace):
     _validate_vm_create_nics(cmd, namespace)
     _validate_vm_vmss_accelerated_networking(cmd.cli_ctx, namespace)
     _validate_vm_vmss_create_auth(namespace)
+    validate_proximity_placement_group(cmd, namespace)
+
     if namespace.secrets:
         _validate_secrets(namespace.secrets, namespace.os_type)
     if namespace.license_type and namespace.os_type.lower() != 'windows':
@@ -1212,6 +1229,7 @@ def process_vmss_create_namespace(cmd, namespace):
     _validate_vm_vmss_accelerated_networking(cmd.cli_ctx, namespace)
     _validate_vm_vmss_create_auth(namespace)
     _validate_vm_vmss_msi(cmd, namespace)
+    validate_proximity_placement_group(cmd, namespace)
 
     if namespace.secrets:
         _validate_secrets(namespace.secrets, namespace.os_type)
@@ -1220,10 +1238,17 @@ def process_vmss_create_namespace(cmd, namespace):
         raise CLIError('usage error: --license-type is only applicable on Windows VM scaleset')
 
     if not namespace.public_ip_per_vm and namespace.vm_domain_name:
-        raise CLIError('Usage error: --vm-domain-name can only be used when --public-ip-per-vm is enabled')
+        raise CLIError('usage error: --vm-domain-name can only be used when --public-ip-per-vm is enabled')
 
     if namespace.eviction_policy and not namespace.priority:
-        raise CLIError('Usage error: --priority PRIORITY [--eviction-policy POLICY]')
+        raise CLIError('usage error: --priority PRIORITY [--eviction-policy POLICY]')
+
+
+def validate_vmss_update_namespace(cmd, namespace):  # pylint: disable=unused-argument
+    if not namespace.instance_id:
+        if namespace.protect_from_scale_in is not None or namespace.protect_from_scale_set_actions is not None:
+            raise CLIError("usage error: protection policies can only be applied to VM instances within a VMSS."
+                           " Please use --instance-id to specify a VM instance")
 # endregion